Read from response `body` instead of `body_io` in endpoint methods
This allows flexibility in test mocks; it allow mocks to return
either `body` or `body_io`.

Previously, mocks could only return `body_io`. This is problematic
if you need to hit the same endpoint twice in the same spec, since
the IO is consumed completely when read the first time, resulting
in error for subsequent reads.